src/metahuman-speechserver/speechserver/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dotmap==1.3.30
elevenlabs==1.2.2
executing==2.0.1
FLAML==2.1.2
Flask==3.0.3
Flask-Cors==4.0.1
google==3.0.0
google-api-core==2.19.0
google-auth==2.30.0
google-cloud==0.34.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
numpy==1.26.4
openai==1.33.0
packaging==24.1
parso==0.8.4
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen==0.2.28
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
pywin32==306
regex==2024.5.15
requests==2.32.3
rsa==4.9
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
SpeechRecognition==3.10.4
stack-data==0.6.3
termcolor==2.4.0
tiktoken==0.7.0
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wexpect==4.0.0
